Don't end stdin in node.js
For some reason this helper module was ending the stdin stream. This seems to have no affect on the specs whether this happens or not, but when it does this there is a random, but highly likely chance that it will exit with a code of 3221225477, which translates into an attempt to access memory in an improper fashion.